refactor(calculator): tableRow 타입 삭제
