+++ app/src/Calculator/components/CalculatorInputContentCardView/index.tsx
@@ -1,11 +1,11 @@
-import type { TableRowList } from '@/Calculator/types/OrderItemButtonGroup';
+import type { Subject } from '@/Calculator/types/Subject';
 
 import { subjectListAtom } from '@/Calculator/atoms/subjectListAtom';
 import { useAtomValue } from 'jotai';
 import { CalculatorInputContentCardList } from './CalculatorInputContentCardList';
 
 export type CalculatorInputContentCardViewProps = {
-  onSubjectListChange: (subjectList: TableRowList[]) => void;
+  onSubjectListChange: (subjectList: Subject[]) => void;
   onInputChange: React.ChangeEventHandler<HTMLInputElement>;
   onCheckboxChange: (index: number) => void; // FIXME: Change to React.ChangeEventHandler<HTMLInputElement> when refactoring
 };
